+++ windows/flutter/generated_plugins.cmake
@@ -5,7 +5,9 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   file_selector_windows
   flutter_secure_storage_windows
+  pdfx
   printing
+  record_windows
   sqlite3_flutter_libs
 )